requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
-e git+https://github.com/rense/django-rest-camel#egg=django-rest-camel
blessings==1.6.1          # via curtsies
bpython==0.17.1
certifi==2018.4.16        # via requests
chardet==3.0.4            # via requests
curtsies==0.3.0           # via bpython
django-extensions==2.0.7
django-js-asset==1.1.0    # via django-mptt
django-mptt==0.9.0
django==2.0.4
djangorestframework-jwt==1.11.0
djangorestframework==3.8.2
greenlet==0.4.13          # via bpython
idna==2.6                 # via requests
invoke==0.22.1
psycopg2==2.7.4
pygments==2.2.0           # via bpython
pyjwt==1.6.1              # via djangorestframework-jwt
pytz==2018.4              # via django
requests==2.18.4          # via bpython
six==1.11.0               # via bpython, django-extensions
typing==3.6.4             # via curtsies
urllib3==1.22             # via requests
wcwidth==0.1.7            # via curtsies
werkzeug==0.14.1
